requirements.txt
# Telegram client
pyrogram>=2.0.106
tgcrypto>=1.2.5

# STT
faster-whisper>=1.0.0
numpy>=1.24.0

# LLM
anthropic>=0.40.0

# TTS
elevenlabs>=1.0.0

# Audio processing
scipy>=1.11.0
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
